requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Faker==4.1.1
idna==2.10
importlib-metadata==1.7.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
